requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
ragas>=0.1.0
langsmith>=0.1.0
